refactor: move trackUser to shared utility
