+++ utils/tracking.test.ts
@@ -1,0 +1,42 @@
+import { trackUser } from './tracking';
+
+describe('trackUser', () => {
+  beforeEach(() => {
+    vi.restoreAllMocks();
+  });
+
+  it('uses navigator.sendBeacon when available', () => {
+    const sendBeaconMock = vi.fn();
+
+    Object.defineProperty(navigator, 'sendBeacon', {
+      value: sendBeaconMock,
+      configurable: true,
+    });
+
+    trackUser('testuser');
+
+    expect(sendBeaconMock).toHaveBeenCalledTimes(1);
+    expect(sendBeaconMock).toHaveBeenCalledWith(
+      '/api/track-user',
+      expect.any(Blob)
+    );
+  });
+
+  it('falls back to fetch when sendBeacon is not available', () => {
+    Object.defineProperty(navigator, 'sendBeacon', {
+      value: undefined,
+      configurable: true,
+    });
+
+    const fetchMock = vi.fn().mockResolvedValue({});
+    global.fetch = fetchMock;
+
+    trackUser('testuser');
+
+    expect(fetchMock).toHaveBeenCalledWith('/api/track-user', {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify({ username: 'testuser' }),
+    });
+  });
+});
